+++ src/main/java/techcourse/myblog/web/MainController.java
@@ -1,0 +1,28 @@
+package techcourse.myblog.web;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import techcourse.myblog.domain.ArticleRepository;
+
+@Controller
+public class MainController {
+    private final ArticleRepository articleRepository;
+
+    @Autowired
+    public MainController(ArticleRepository articleRepository) {
+        this.articleRepository = articleRepository;
+    }
+
+    @GetMapping("/")
+    public String index(Model model) {
+        model.addAttribute("articles", articleRepository.findAll());
+        return "index";
+    }
+
+    @GetMapping("/writing")
+    public String createArticleForm() {
+        return "article-edit";
+    }
+}
