+++ src/main/resources/templates/article-edit.html
@@ -21,7 +21,7 @@
                 <div class="height-300px vertical-center pdd-horizon-50">
                     <div class="width-80">
                         <input type="hidden" name="id" th:if="${article} != null" th:value="${article.id}"/>
-                        <input type="text" name="title" th:value="${article} == null ? '' : ${article.title}" class="text-center bg-transparent no-border font-size-30 width-60 mx-auto d-block text-white border bottom" autocomplete="off" placeholder="제목을 입력해주세요">
+                        <input type="text" name="title" th:value="${article} == null ? '' : ${article.title}" class="text-center bg-transparent no-border font-size-30 width-60 mx-auto d-block text-white border bottom" autocomplete="off" placeholder="제목을 입력해주세요" autofocus>
                         <input type="text" name="coverUrl" th:value="${article} == null ? '' : ${article.coverUrl}" class="text-center bg-transparent no-border font-size-16 width-30 mx-auto d-block text-light text-white border bottom pdd-top-20"  placeholder="배경 url을 입력해주세요" autocomplete="off">
                     </div>
                 </div>
