+++ src/test/java/techcourse/myblog/web/MainControllerTests.java
@@ -1,0 +1,36 @@
+package techcourse.myblog.web;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpMethod;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.web.reactive.server.WebTestClient;
+
+@AutoConfigureWebTestClient
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class MainControllerTests {
+    @Autowired
+    private WebTestClient webTestClient;
+
+    @Test
+    void index_페이지_조회() {
+        statusIsOk(HttpMethod.GET, "/");
+    }
+
+    @Test
+    void 게시글_작성_페이지_확인() {
+        statusIsOk(HttpMethod.GET, "/writing");
+    }
+
+    private void statusIsOk(HttpMethod httpMethod, String uri) {
+        webTestClient.method(httpMethod)
+                .uri(uri)
+                .exchange()
+                .expectStatus()
+                .isOk();
+    }
+}
